pyproject.toml
[project]
name = "telegram-shell"
version = "0.1.0"
description = "Personal Telegram bot that runs shell commands in a persistent session"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "pexpect>=4.9",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[tool.uv]
package = false
